Move network package to lib
