+++ solution/parser/ecfr-parser/eregs/eregs.go
@@ -9,9 +9,10 @@ import (
 	"path"
 	"time"
 
-	"github.com/cmsgov/cmcs-eregulations/ecfr-parser/ecfr"
-	"github.com/cmsgov/cmcs-eregulations/network"
+	"github.com/cmsgov/cmcs-eregulations/ecfr-parser/ecfr"	
 	"github.com/cmsgov/cmcs-eregulations/ecfr-parser/parsexml"
+
+	"github.com/cmsgov/cmcs-eregulations/lib/network"
 
 	log "github.com/sirupsen/logrus"
 )
