+++ solution/parser/fr-parser/eregs/eregs.go
@@ -9,7 +9,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/cmsgov/cmcs-eregulations/network"
+	"github.com/cmsgov/cmcs-eregulations/lib/network"
 
 	log "github.com/sirupsen/logrus"
 )
