+++ solution/parser/lib/network/go.mod
