+++ solution/parser/lib/network/network.go
